examples/autonomous-portfolio-agent/pyproject.toml
[project]
name = "autonomous-portfolio-agent"
version = "0.1.0"
description = "Autonomous crypto portfolio manager with honeypot protection via PolicyLayer"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "coinbase-agentkit>=0.2.0",
    "coinbase-agentkit-langchain>=0.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "web3>=7.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "colorama>=0.4.6",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100